+++ __tests__/e2e__basic.test.ts
@@ -4,22 +4,20 @@ import { isAsset, isChunk } from '../src/helpers'
 import { ChromeExtensionManifest } from '../src/manifest'
 import { byFileName, getExtPath } from '../__fixtures__/utils'
 
-const { default: config } = require(getExtPath(
-  'basic/rollup.config.js',
-))
+const { default: config } = require(getExtPath('basic/rollup.config.js'))
 
 let bundle: RollupBuild
 let output: [OutputChunk, ...(OutputChunk | OutputAsset)[]]
 let ready: Promise<void>
 beforeAll(async () => {
+  // TODO: make live bundle imports stub
+  // __fixtures__/bundle-imports-stub/rollup.config.js
   try {
     bundle = await rollup(config)
 
-    ready = bundle
-      .generate(config.output)
-      .then(({ output: o }) => {
-        output = o
-      })
+    ready = bundle.generate(config.output).then(({ output: o }) => {
+      output = o
+    })
 
     if (!process.env.JEST_WATCH) {
       await writeJSON(getExtPath('basic-build.json'), bundle, {
@@ -62,45 +60,23 @@ test(
 
     // 17 assets + 2 wrapper scripts
     expect(output.find(byFileName('asset.js'))).toBeDefined()
-    expect(
-      output.find(byFileName('popup/popup.html')),
-    ).toBeDefined()
-    expect(
-      output.find(byFileName('devtools/devtools.html')),
-    ).toBeDefined()
-    expect(
-      output.find(byFileName('images/icon-main-16.png')),
-    ).toBeDefined()
-    expect(
-      output.find(byFileName('images/icon-main-48.png')),
-    ).toBeDefined()
-    expect(
-      output.find(byFileName('images/icon-main-128.png')),
-    ).toBeDefined()
-    expect(
-      output.find(byFileName('images/favicon.ico')),
-    ).toBeDefined()
-    expect(
-      output.find(byFileName('images/favicon.png')),
-    ).toBeDefined()
+    expect(output.find(byFileName('popup/popup.html'))).toBeDefined()
+    expect(output.find(byFileName('devtools/devtools.html'))).toBeDefined()
+    expect(output.find(byFileName('images/icon-main-16.png'))).toBeDefined()
+    expect(output.find(byFileName('images/icon-main-48.png'))).toBeDefined()
+    expect(output.find(byFileName('images/icon-main-128.png'))).toBeDefined()
+    expect(output.find(byFileName('images/favicon.ico'))).toBeDefined()
+    expect(output.find(byFileName('images/favicon.png'))).toBeDefined()
     expect(output.find(byFileName('options.html'))).toBeDefined()
     expect(output.find(byFileName('options.css'))).toBeDefined()
     expect(output.find(byFileName('content.css'))).toBeDefined()
     expect(output.find(byFileName('options.png'))).toBeDefined()
     expect(output.find(byFileName('options.jpg'))).toBeDefined()
-    expect(
-      output.find(byFileName('manifest.json')),
-    ).toBeDefined()
+    expect(output.find(byFileName('manifest.json'))).toBeDefined()
 
-    expect(
-      output.find(byFileName('fonts/NotoSans-Light.ttf')),
-    ).toBeDefined()
-    expect(
-      output.find(byFileName('fonts/NotoSans-Black.ttf')),
-    ).toBeDefined()
-    expect(
-      output.find(byFileName('fonts/Missaali-Regular.otf')),
-    ).toBeDefined()
+    expect(output.find(byFileName('fonts/NotoSans-Light.ttf'))).toBeDefined()
+    expect(output.find(byFileName('fonts/NotoSans-Black.ttf'))).toBeDefined()
+    expect(output.find(byFileName('fonts/Missaali-Regular.otf'))).toBeDefined()
   },
   5 * 60 * 1000,
 )
@@ -109,19 +85,12 @@ test('Includes content script imports in web_accessible_resources', async () => 
   expect(ready).toBeDefined()
   await ready
 
-  const manifestAsset = output.find(
-    byFileName('manifest.json'),
-  ) as OutputAsset
+  const manifestAsset = output.find(byFileName('manifest.json')) as OutputAsset
   const manifestSource = manifestAsset.source as string
-  const manifest: ChromeExtensionManifest = JSON.parse(
-    manifestSource,
-  )
+  const manifest: ChromeExtensionManifest = JSON.parse(manifestSource)
 
   expect(manifest).toMatchObject({
-    web_accessible_resources: expect.arrayContaining([
-      'content.js',
-      expect.stringMatching(/imported-.+?\.js/),
-    ]),
+    web_accessible_resources: expect.arrayContaining(['content.js', expect.stringMatching(/imported-.+?\.js/)]),
   })
 })
 
@@ -129,17 +98,12 @@ test('Includes content_security_policy untouched', async () => {
   expect(ready).toBeDefined()
   await ready
 
-  const manifestAsset = output.find(
-    byFileName('manifest.json'),
-  ) as OutputAsset
+  const manifestAsset = output.find(byFileName('manifest.json')) as OutputAsset
   const manifestSource = manifestAsset.source as string
-  const manifest: ChromeExtensionManifest = JSON.parse(
-    manifestSource,
-  )
+  const manifest: ChromeExtensionManifest = JSON.parse(manifestSource)
 
   expect(manifest).toMatchObject({
-    content_security_policy:
-      "script-src 'self'; object-src 'self'",
+    content_security_policy: "script-src 'self'; object-src 'self'",
   })
 })
