+++ src/manifest-input/browser/__mocks__/placeholders.ts
@@ -1,0 +1,3 @@
+export const eventPaths = JSON.parse('["runtime.onMessage","runtime.onInstalled"]') as string[]
+export const importPath = JSON.parse('"import/path"') as string
+export const delayLength = JSON.parse('1000') as number
