+++ src/manifest-input/browser/contentScriptWrapper.ts
@@ -1,0 +1,1 @@
+// TODO: write content script wrapper
