+++ src/manifest-input/browser/defer.ts
@@ -1,0 +1,20 @@
+export function defer<T>() {
+  let resolve: (value?: T) => void
+  let reject: (error: any) => void
+  const promise = new Promise<T>((res, rej) => {
+    resolve = res
+    reject = rej
+  })
+
+  return Object.assign(promise, {
+    resolve(value?: T) {
+      resolve!(value)
+    },
+    reject(error: any) {
+      reject!(error)
+    },
+  }) as Promise<T> & {
+    resolve: typeof resolve
+    reject: typeof reject
+  }
+}
