+++ src/manifest-input/browser/placeholders.ts
@@ -1,0 +1,3 @@
+export const eventPaths = JSON.parse('%EVENTS%') as string[]
+export const importPath = JSON.parse('%PATH%') as string
+export const delayLength = JSON.parse('%DELAY%') as number
